GitHub Actions; Refactor tests (#524)
* refactor test structure

* first try github action

* u+x

* fix

* fix

* add ci_pr.yml

* fix docs

* remove docs test; add repo size test

* add deploy script

* require petabtests>=0.0.0a6

* add pysb

* split nb tests; fix something

* split up base tests

* fix

* fix

* fix ipopt

* tidy up

* fix

* fix

* separate petab

* fixup

* shutup atol

* update badges

* add docs test

* fix

* fix

* fix

* tidy up

* address review comments

* amici -> 0.11.9
