--- .travis.yml
@@ -1,90 +1,0 @@
-language: python
-python:
-  - "3.8"
-
-os:
-  - linux
-dist:
-  - bionic
-compiler:
-  - gcc
-
-# install dependencies
-
-addons:
-  apt:
-    update: true
-    packages:
-      - build-essential
-      - coinor-libipopt1v5
-      - coinor-libipopt-dev
-      - gfortran
-      - lcov
-      - libatlas-base-dev
-      - libhdf5-serial-dev
-      - pandoc
-      - pkg-config
-      - python-dev
-      - swig3.0
-      - zlib1g-dev
-
-env:
-  - BNGPATH=/home/travis/build/ICB-DCM/pyPESTO/BioNetGen-2.3.2
-
-install:
-  - mkdir -p ~/bin/ && ln -s /usr/bin/swig3.0 ~/bin/swig && export PATH=~/bin/:$PATH
-  - pip3 install -r ./.travis_pip_reqs.txt
-  - pip3 install .[ipopt,dlib,pyswarm,pymc3,cmaes,nlopt,fides]
-  - pip3 install https://github.com/petab-dev/petab/archive/develop.zip
-  - pip3 install -e git+https://github.com/pysb/pysb.git@c434f6ab98301beee1bf9d2a5093f0c79da78824#egg=pysb
-  - pip3 install -e
-      git+https://github.com/icb-dcm/amici.git@develop#egg=amici\&subdirectory=python/sdist
-  - if [ "$TRAVIS_PULL_REQUEST" = "false" ]; then
-      git clone --depth 1
-        https://github.com/petab-dev/petab_test_suite tmp/petab_test_suite;
-      pip3 install -e tmp/petab_test_suite;
-      wget -q -O bionetgen.tar.gz https://bintray.com/jczech/bionetgen/download_file?file_path=BioNetGen-2.3.2-linux.tar.gz && tar -xzf bionetgen.tar.gz;
-    fi
-  - if [ "$TRAVIS_PULL_REQUEST" != "false" ]; then
-      pip install -r .rtd_pip_reqs.txt;
-    fi
-
-# run tests
-script:
-  - if [ "$TRAVIS_PULL_REQUEST" = "false" ]; then
-      ./run_flake8.sh;
-    fi
-  - if [ "$TRAVIS_PULL_REQUEST" = "false" ]; then
-      travis_wait 20 python3 -m pytest --cov=pypesto ./test/test_*;
-    fi
-  - if [ "$TRAVIS_PULL_REQUEST" = "false" ]; then
-      travis_wait 20 xvfb-run -a python3 -m pytest
-        --cov=pypesto --cov-append ./test/visualize/test_*;
-    fi
-  - if [ "$TRAVIS_PULL_REQUEST" != "false" ]; then
-      sphinx-build -W -b html -d doc/_build/html doc/ doc/_build/html;
-    fi
-  - if [ "$TRAVIS_PULL_REQUEST" != "false" ]; then
-      travis_wait 20 test/run_notebook.sh doc/example/;
-    fi
-  - if [ "$TRAVIS_PULL_REQUEST" = "false" ]; then
-      coverage xml;
-    fi
-  - if [ "$TRAVIS_PULL_REQUEST" = "false" ]; then
-      ./test/check_git_size.sh;
-    fi
-
-after_success:
-  - bash <(curl -s https://codecov.io/bash)
-
-deploy:
-  provider: pypi
-  username: $TWINE_USERNAME
-  password: $TWINE_PASSWORD
-  on:
-    tags: true
-  skip_existing: true
-
-# cache dependencies
-
-cache: pip
