--- .travis_pip_reqs.txt
@@ -1,17 +1,0 @@
-setuptools
-wheel
-pkgconfig
-jupyter
-pytest
-pytest-cov
-coverage
-flake8
-flake8-bugbear
-flake8-bandit
-flake8-print
-flake8-comprehensions
-gitpython
-sphinx
-nbsphinx
-sphinx_rtd_theme
-cython
