+++ setup.py
@@ -30,12 +30,13 @@ setup(name='pypesto',
                         'seaborn>=0.10.0',
                         'cloudpickle>=0.7.0',
                         'h5py>=2.10.0',
-                        'tqdm>=4.46.0'],
+                        'tqdm>=4.46.0',
+                        'gitpython>=3.1.7'],
       tests_require=['pytest>=5.4.2',
                      'flake8>=3.7.2',
                      'gitpython>=3.1.2',
                      'petabtests'],
-      extras_require={'amici': ['amici>=0.11.8'],
+      extras_require={'amici': ['amici>=0.11.9'],
                       'petab': ['petab>=0.1.11'],
                       'ipopt': ['ipopt>=0.1.9'],
                       'dlib': ['dlib>=19.19.0'],
@@ -46,6 +47,18 @@ setup(name='pypesto',
                       'pymc3': ['arviz>=0.8.1, <0.9.0',
                                 'theano>=1.0.4',
                                 'packaging>=20.0',
-                                'pymc3>=3.8, <3.9.2']},
+                                'pymc3>=3.8, <3.9.2'],
+                      'test': ['pytest>=5.4.3',
+                               'pytest-cov>=2.10.0'],
+                      'docs': ['sphinx>=3.1.0',
+                               'nbsphinx>=0.7.0',
+                               'nbconvert>=5.6.1',
+                               'sphinx-rtd-theme>=0.4.3'],
+                      'quality': ['flake8>=3.8.3',
+                                  'flake8-bugbear>=20.1.4'
+                                  'flake8-bandit>=2.1.2'
+                                  'flake8-print>=3.1.4'
+                                  'flake8-comprehensions>=3.2.3'],
+                      },
       python_requires='>=3.6',
       )
