+++ test/petab/__init__.py
@@ -1,0 +1,1 @@
+"""PEtab and AMICI related tests."""
