+++ test/petab/test_petab_import_pysb.py
@@ -3,22 +3,27 @@ This is for testing the petab import.
 """
 
 import os
-
 import numpy as np
-import petabtests
 import yaml
+import logging
 
 import pypesto.optimize as optimize
 from pypesto.petab.pysb_importer import PetabImporterPysb
 # must import after previous, otherwise circular import issues :(
 from amici.petab_import_pysb import PysbPetabProblem
+import petabtests
+
+# In CI, bionetgen is install here
+BNGPATH = os.path.abspath(
+    os.path.join(os.path.dirname(__file__), '..', '..', 'BioNetGen-2.3.2'))
+if 'BNGPATH' not in os.environ:
+    logging.warning(f"Env var BNGPATH was not set. Setting to {BNGPATH}")
+    os.environ['BNGPATH'] = BNGPATH
 
 
 def test_petab_pysb_optimization():
     test_case = '0001'
-    repo_base = os.path.join(os.path.dirname(os.path.abspath(__file__)), "..")
-    test_case_dir = os.path.join(repo_base, 'tmp', 'petab_test_suite',
-                                 'cases', 'pysb', test_case)
+    test_case_dir = os.path.join(petabtests.PYSB_DIR, test_case)
     petab_yaml = os.path.join(test_case_dir, f'_{test_case}.yaml')
     soution_yaml = os.path.join(test_case_dir, f'_{test_case}_solution.yaml')
