+++ test/profile/__init__.py
@@ -1,0 +1,1 @@
+"""Profiling tests."""
